refactor visitor
